Revert "GEODE-7309: uplift lucene to 7.1.0 (#6351)" (#6903)
This reverts commit 68629356f561a932f5dfbace70b01d9971a42473.

+++ geode-lucene/src/main/java/org/apache/geode/cache/lucene/internal/LuceneIndexStats.java
@@ -120,7 +120,7 @@ public class LuceneIndexStats {
   /**
    * @param start the timestamp taken when the operation started
    */
-  public void endRepositoryQuery(long start, final long totalHits) {
+  public void endRepositoryQuery(long start, final int totalHits) {
     stats.incLong(repositoryQueryExecutionTimeId, getStatTime() - start);
     stats.incInt(repositoryQueryExecutionsInProgressId, -1);
     stats.incInt(repositoryQueryExecutionsId, 1);
